gateway yml 수정
