+++ gateway-service/src/main/resources/application.yml
@@ -12,7 +12,6 @@ spring:
             allowed-origins:
               - "http://localhost:5173" #사용자 프론트 서버
               - "https://samubozo.co.kr" #AWS 정적 서버
-
             allowed-methods:
               - GET
               - POST
